type-coercion

+++ server/test/org/sogive/data/charity/MoneyTest.java
@@ -11,7 +11,7 @@ public class MoneyTest {
 		Money two = Money.pound(2);
 		com.goodloop.data.Money three = com.goodloop.data.Money.pound(3);
 		Money five = two.plus(three);
-		assert five.getValue() == 5 : five;
+		assert five.getValue().doubleValue() == 5 : five;
 	}
 
 	@Test
@@ -19,7 +19,7 @@ public class MoneyTest {
 		Money three = Money.pound(3);
 		com.goodloop.data.Money two = com.goodloop.data.Money.pound(2);
 		Money one = three.minus(two);
-		assert one.getValue() == 1 : one;
+		assert one.getValue().doubleValue() == 1 : one;
 	}
 
 }
